Clean ui.stories.js
